fix: Moved fpmm/trading functions inside one file,renamed utils to math,fixed tests and made new http file
